第 22 轮回归成果：权威 md 1118→1103（累计 3324→1103）
- 逐轮 commit 制度建立（每轮结束提交一次）
- 第 22 轮：statics 85→75、guild 354→349；auction/point 各剩 1
- 类粒度子分解持续：dbmw×5、manager×3 子 agent 组（分组记录见 round22_groups/）
- 按新规矩（禁止硬套 asm）纯 C++ 形态修复
- 权威口径：先 gen_report_manifest 再 gen_function_md 全量重生成

+++ dnf_decompile/source/DNFServer/GameServer/Monitor/DNFTowerRank.cpp
@@ -112,7 +112,7 @@ void CTowerRank::reset()
 void CTowerRank::registRank(unsigned int floor, unsigned int job, unsigned int score,
                             const stDeathTowerRecordMemberInfo_t* records)
 {
-    for (unsigned int i = 0; i < floor; i++)
+    for (int i = 0; i < (int)floor; i++)
     {
         registCharacRank(floor, records[i].m_name, job, score);
     }
